requirements.txt
requests
python-telegram-bot
pydantic